update

Created-by: msprofDeveloper

Match-id-b18a71bc0b21438a769fba96e8110e9ecadf8185

+++ collector/dvvp/adx/wrapper/hdc_api.cpp
@@ -375,7 +375,7 @@ static hdcError_t HdcWritePackage(HDC_SESSION session, DataSendMsg dataSendMsg,
     hdcError_t hdcError = DRV_ERROR_NONE;
     unsigned int totalLen = static_cast<uint32_t>(dataSendMsg.bufLen);
     unsigned int reservedLen = totalLen;
-    int sendLen = 0;
+    unsigned int sendLen = 0;
     unsigned int timeout = 0;
     IdeSendBuffT buf = dataSendMsg.buf;
 
@@ -389,14 +389,15 @@ static hdcError_t HdcWritePackage(HDC_SESSION session, DataSendMsg dataSendMsg,
             sendLen = dataSendMsg.maxSendLen;
             packet->isLast = static_cast<int8_t>(IdeLastPacket::IDE_NOT_LAST_PACK);
         } else {
-            sendLen = static_cast<int32_t>(reservedLen);
+            sendLen = reservedLen;
             packet->isLast = static_cast<int8_t>(IdeLastPacket::IDE_LAST_PACK);
         }
         packet->len = sendLen;
         packet->type = IDE_DAEMON_LITTLE_PACKAGE;
 
         errno_t ret = memcpy_s(packet->value, dataSendMsg.maxSendLen,
-                               static_cast<IdeU8Pt>(const_cast<IdeBuffT>(buf)) + (totalLen - reservedLen), sendLen);
+                               static_cast<IdeU8Pt>(const_cast<IdeBuffT>(buf)) + (totalLen - reservedLen),
+                               static_cast<int32_t>(sendLen));
         IDE_CTRL_VALUE_FAILED(ret == EOK, return DRV_ERROR_INVALID_VALUE, "memory copy failed");
 
         // add buffer to hdc message
@@ -664,7 +665,7 @@ int IdeCreatePacket(CmdClassT type, IdeString value,
     IDE_CTRL_VALUE_FAILED(sendBuf != nullptr, return IDE_DAEMON_ERROR, "malloc memory failed");
     IdeTlvReq req = (IdeTlvReq)sendBuf;
     req->dev_id = 0;
-    req->len = valueLen;
+    req->len = static_cast<int32_t>(valueLen);
     req->type = type;
     errno_t err = memcpy_s(req->value, mallocValueLen, value, valueLen);
     if (err != EOK) {
